feat(organization): changed files to their perspective folder
