+++ include/frontend/parser/printer/nodes/expressions/print_assignment.h
@@ -1,9 +1,8 @@
 #ifndef PRINT_ASSIGNMENT_H
 #define PRINT_ASSIGNMENT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
-
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_assignment(const AstNode *node, int depth, VisitedNodes *visited);
