+++ include/frontend/parser/printer/nodes/expressions/print_binary_expr.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_BINARY_EXPR_H
 #define PRINT_BINARY_EXPR_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_binary_expr(const AstNode *node, int depth, VisitedNodes *visited);
