+++ include/frontend/parser/printer/nodes/expressions/print_identifier.h
@@ -1,7 +1,7 @@
 #ifndef PRINT_IDENTIFIER_H
 #define PRINT_IDENTIFIER_H
 
-#include "../../../ast/definitions.h"
+#include "frontend/ast/definitions.h"
 
 void print_identifier(const AstNode *node, int depth);
