+++ include/frontend/parser/printer/nodes/expressions/print_logical_not.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_LOGICAL_NOT_H
 #define PRINT_LOGICAL_NOT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_logical_not(const AstNode *node, int depth, VisitedNodes *visited);
