+++ include/frontend/parser/printer/nodes/expressions/print_numeric_literal.h
@@ -1,7 +1,7 @@
 #ifndef PRINT_NUMERIC_LITERAL_H
 #define PRINT_NUMERIC_LITERAL_H
 
-#include "../../../ast/definitions.h"
+#include "frontend/ast/definitions.h"
 
 void print_numeric_literal(const AstNode *node, int depth);
