+++ include/frontend/parser/printer/nodes/expressions/print_object.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_OBJECT_H
 #define PRINT_OBJECT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_object(const AstNode *node, int depth, VisitedNodes *visited);
