+++ include/frontend/parser/printer/nodes/expressions/print_pre_decrement.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_PRE_DECREMENT_H
 #define PRINT_PRE_DECREMENT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_pre_decrement(const AstNode *node, int depth, VisitedNodes *visited);
