+++ include/frontend/parser/printer/nodes/expressions/print_pre_increment.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_PRE_INCREMENT_H
 #define PRINT_PRE_INCREMENT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_pre_increment(const AstNode *node, int depth, VisitedNodes *visited);
