+++ include/frontend/parser/printer/nodes/expressions/print_property.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_PROPERTY_H
 #define PRINT_PROPERTY_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_property(const AstNode *node, int depth, VisitedNodes *visited);
