+++ include/frontend/parser/printer/nodes/expressions/print_unary_bitwise_not.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_UNARY_BITWISE_NOT_H
 #define PRINT_UNARY_BITWISE_NOT_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_unary_bitwise_not(const AstNode *node, int depth, VisitedNodes *visited);
