+++ include/frontend/parser/printer/nodes/print_program.h
@@ -1,8 +1,8 @@
 #ifndef PRINT_PROGRAM_H
 #define PRINT_PROGRAM_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_program(const AstNode *node, int depth, VisitedNodes *visited);
