+++ include/frontend/parser/printer/nodes/statements/print_function.h
@@ -1,9 +1,8 @@
 #ifndef PRINT_FUNCTION_H
 #define PRINT_FUNCTION_H
 
-#include "../../../ast/definitions.h"
-#include "../visited.h"
-
+#include "frontend/ast/definitions.h"
+#include "frontend/parser/printer/visited.h"
 
 void print_function(const AstNode *node, int depth, VisitedNodes *visited);
