+++ include/frontend/parser/printer/nodes/statements/print_import.h
@@ -1,7 +1,7 @@
 #ifndef PRINT_IMPORT_H
 #define PRINT_IMPORT_H
 
-#include "../../../ast/definitions.h"
+#include "frontend/ast/definitions.h"
 
 void print_import(const AstNode *node, int depth);
