+++ include/frontend/parser/printer/nodes/statements/print_package.h
@@ -1,7 +1,7 @@
 #ifndef PRINT_PACKAGE_H
 #define PRINT_PACKAGE_H
 
-#include "../../../ast/definitions.h"
+#include "frontend/ast/definitions.h"
 
 void print_package(const AstNode *node, int depth);
