+++ src/frontend/parser/printer/nodes/expressions/print_property.c
@@ -1,4 +1,4 @@
-#include "frontend/parser/printer/nodes/print_property.h"
+#include "frontend/parser/printer/nodes/expressions/print_property.h"
 #include "frontend/parser/printer/visited.h"
 #include "frontend/ast/definitions.h"
 #include "frontend/parser/printer/print_indent.h"
