+++ src/frontend/parser/printer/nodes/expressions/print_unary_bitwise_not.c
@@ -1,4 +1,4 @@
-#include "frontend/parser/printer/nodes/print_unary_bitwise_not.h"
+#include "frontend/parser/printer/nodes/expressions/print_unary_bitwise_not.h"
 #include "frontend/parser/printer/visited.h"
 #include "frontend/ast/definitions.h"
 #include "frontend/parser/printer/print_indent.h"
