+++ src/frontend/parser/printer/print_ast.c
@@ -5,22 +5,22 @@
 #include "frontend/parser/printer/print_indent.h"
 #include "frontend/parser/printer/visited.h"
 #include "frontend/parser/printer/nodes/print_program.h"
-#include "frontend/parser/printer/nodes/print_assignment.h"
-#include "frontend/parser/printer/nodes/print_binary_expr.h"
-#include "frontend/parser/printer/nodes/print_identifier.h"
-#include "frontend/parser/printer/nodes/print_property.h"
-#include "frontend/parser/printer/nodes/print_package.h"
-#include "frontend/parser/printer/nodes/print_object.h"
-#include "frontend/parser/printer/nodes/print_numeric_literal.h"
-#include "frontend/parser/printer/nodes/print_import.h"
-#include "frontend/parser/printer/nodes/print_logical_not.h"
-#include "frontend/parser/printer/nodes/print_unary_minus.h"
-#include "frontend/parser/printer/nodes/print_unary_bitwise_not.h"
-#include "frontend/parser/printer/nodes/print_pre_decrement.h"
-#include "frontend/parser/printer/nodes/print_pre_increment.h"
-#include "frontend/parser/printer/nodes/print_variable.h"
-#include "frontend/parser/printer/nodes/print_function.h"
-#include "frontend/parser/printer/nodes/print_for.h"
+#include "frontend/parser/printer/nodes/expressions/print_assignment.h"
+#include "frontend/parser/printer/nodes/expressions/print_binary_expr.h"
+#include "frontend/parser/printer/nodes/expressions/print_identifier.h"
+#include "frontend/parser/printer/nodes/expressions/print_property.h"
+#include "frontend/parser/printer/nodes/expressions/print_object.h"
+#include "frontend/parser/printer/nodes/expressions/print_numeric_literal.h"
+#include "frontend/parser/printer/nodes/expressions/print_logical_not.h"
+#include "frontend/parser/printer/nodes/expressions/print_unary_minus.h"
+#include "frontend/parser/printer/nodes/expressions/print_unary_bitwise_not.h"
+#include "frontend/parser/printer/nodes/expressions/print_pre_decrement.h"
+#include "frontend/parser/printer/nodes/expressions/print_pre_increment.h"
+#include "frontend/parser/printer/nodes/statements/print_import.h"
+#include "frontend/parser/printer/nodes/statements/print_package.h"
+#include "frontend/parser/printer/nodes/statements/print_variable.h"
+#include "frontend/parser/printer/nodes/statements/print_function.h"
+#include "frontend/parser/printer/nodes/statements/print_for.h"
 #include "frontend/parser/printer/print_ast.h"
 #include "frontend/parser/printer/visited.h"
